Backend Neo4j/requirements.txt
# Core scientific stack
numpy==1.26.4
pandas==2.1.4
scipy==1.10.1
scikit-learn==1.3.2
joblib==1.3.2

# NLP + text preprocessing
ftfy==6.1.1
langdetect==1.0.9
Unidecode==1.3.8
sentence-transformers==2.6.1

# PDF + parsing
pypdf==4.2.0

# LangChain ecosystem (consistent versions)
langchain==0.2.12
langchain-core>=0.2.27,<0.3.0
langchain-community==0.2.11
langchain-text-splitters==0.2.2
langchain-chroma==0.1.4
langchain-huggingface==0.0.3

# Vector stores / embeddings
chromadb==0.5.0
pyarrow==15.0.2

# Graph DB
neo4j==5.20.0

# ML models
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Pydantic (v1.x for FastAPI + training code compatibility)
pydantic==1.10.22

# Utilities
tqdm==4.66.4
matplotlib==3.8.2
seaborn==0.13.2

# API stack
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
